pyproject.toml
[project]
name = "word-vector-embedding"
version = "0.1.0"
description = "Display word embedding vector"
requires-python = ">=3.11"
dependencies = [
    "streamlit",
    "scikit-learn",
    "gensim",
    "sentence-transformers",
    "python-docx",
    "pandas",
    "plotly",
    "pdfplumber",
    "matplotlib",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"